Isolate out robot interface base class
